[IMP] html_editor: correctly format inserted table
Prior to this commit, when inserting a table in an html_field, it would
use tags such as `thead` which are not supposed by the table plugin when
doing table manipulations.

With this commit, we create an insertion handler which changes any thead
to tbody and applies the `o_table_header` class to the rows so it
visually resembles a thead.

Task-5164382