[IMP] purchase_stock: log activity in PO without backorder
In Sales, if an SO has an incomplete picking without a backorder, there
will be an activity logged in the SO informing about the situation. In
Purchase, in a similar situation, only a note about an updated quantity
is logged. This commit adds activity creation to Purchase as well,
making the behavior consistent between these modules.

Task: 5354522