[FIX] pos_restaurant: update correct product qty on ticket screen
Currently, modifying a loaded order (e.g., changing product quantity) does
not save the changes, and the old order without modifications is shown on
the ticket screen.

Steps to Reproduce:
1) Install the POS Restaurant app with demo data.
2) Open a Restaurant Session.
3) Click on 'Register' to create a Direct Sale.
4) Add a product (e.g., Water with Qty 2) and a customer (e.g., 'Billy Fox').
5) Click on 'Orders' from the navbar and load the recently created order.
6) Update the quantity of Water from 2 to 5 and click on 'Orders' again.

Error:
The quantity of Water is not updated, the order line still shows a quantity
of 2.

Root Cause:
Clicking on 'Orders' calls `syncAllOrders` (see [1]), which is responsible
for updating the orders. However, `this.getPendingOrder()` returns null
in this scenario, causing the orders array to be empty and the function to
return early without syncing.

Fix:
Override `sendOrderToCustomerDisplay`. This method is invoked whenever the
order is updated (e.g., adding lines, changing quantities) to refresh the
display. By calling `addPendingOrder` within this method, we ensure that
the modifications are registered, allowing `syncAllOrders` to correctly
sync the order with the server.

[1]: https://github.com/odoo/odoo/blob/7fd5f90fe8fce4de49ec10c683f92e7a3557981a/addons/point_of_sale/static/src/app/services/pos_store.js#L1485-L1512

opw-5405402