[IMP] account_*,sale_*,l10n_*: add default payment term configuration
Currently, when creating a sales order or an invoice, the payment term (or due
date) defaults to immediate payment.

In practice, most companies use the same payment term for the majority of their
documents. This improvement introduces a default payment term at the company
level, allowing it to be automatically applied on sales orders and invoices.

This avoids having to manually change the payment term each time or define it
explicitly on every partner.

Task-5387239
see https://github.com/odoo/enterprise/pull/103168