[FIX] html_builder: prevent invalid snippet position
Before this commit, the dropzones activation, after clicking on a
category in the block tab and choosing a snippet, would not be correct.

The dropzones would be activated based on a default snippet template
and not on the selected snippet. Therefore, any rules that normally
prevent the snippet from behing drop somewhere would be ignore.

This commit fixes the issue by using the chosen snippet to determine
which dropzones should be activated.

task-5428387