[IMP] mrp_account*: move compute price from BoM action to BOM list view
*: mrp_subcontracting_account, purchase_mrp, sale_mrp, sale_mrp_margin

Currently, the `Compute price from BoM` action is available directly on the
product form. Because it is highly visible and easily accessible, users tend to
trigger it frequently, even when there is no need to recompute the cost.

Moreover, in `AVCO and FIFO` cost configurations, manually updating the product
cost based on the BoM is conceptually incorrect. Triggering this action can lead
to misleading cost values and may cause inventory valuation or accounting
inconsistencies.

This PR moves cost computation actions away from the product form and places
them at the BoM level, which is the actual source of truth for this information.

- The `Compute price from BoM` action has been removed from the product form
to prevent accidental or repetitive usage. A new `Update Product Cost from BoM`
action is made available directly on the BoM list view.
- The `Unit Cost` is now displayed by default in the BoM list view, providing
better visibility into BoM-driven costs.

Overall, this improvement minimizes unintended cost updates, and ensures that
BoM-based cost computation is used only when it is functionally relevant,
resulting in safer and more predictable accounting behavior.

TaskID-5180166