[FIX] html_editor: use start and end in text alignment
Since [1] when text alignment was introduced on the toolbar, `left` and
`right` were used for `text-align` instead of `start` and `end`. This
causes issues in RTL languages, especially given that oriented icons are
already flipped according to the direction.

This commit uses start/end instead of left/right for text alignment.

Steps to reproduce:
- Set the user's language to Arabic
- Go to a "To do" note
- Change the text alingment to left or right

=> The icon in the toolbar showed the opposite

[1]: https://github.com/odoo/odoo/pull/191746

task-5155800