[FIX] mrp: variant synchronisation on BoM report
version : 17.0+e

Steps to reproduce on runbot
------------------
Select a product with several variants and a Bill of Materials (e.g. Stool).

Change the variants order so that their ids are not ordered, this can be done by modifying the default_code for example(e.g. Internal Reference for variant "Color: Green" set to "A").

When accessing the BoM report, you won’t be able to switch to one of the possible variants (in the example the Dark Blue variant).

Why it is happening
------------------
The default variant to be displayed when opening the report is selected in the backend using the product_variant_id field. This field is computed as the first element in product_variant_ids as they are ordered in the model.

We then send this variant’s information to the frontend and a dictionary containing every variant (key= id and value = display_name). In the serialization process, the object is reordered based on the keys. Thus, if the variants were not ordered based on their ids in python, the order will change.

The displayed variant is correct as it has been passed directly but the frontend also computes the currentVariant attribute. This is computed as the first element in the dictionary but in this case, it is not the one that has been selected in the backend, as the order changed.

As a result, you see the report for a variant A but the frontend considers you are on the report for variant B so you cannot switch to variant B as you are supposed to be already on it.

The fix
------------------
I propose to explicitly pass the selected variant to the frontend.

opw-5409493